pyproject.toml
[build-system]
requires = ["build", "wheel", "setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pandapipes"
version = "0.12.post2"  # File format version '__format_version__' is tracked in __init__.py
authors=[
    { name = "Simon Ruben Drauz-Mauel", email = "simon.ruben.drauz-mauel@iee.fraunhofer.de" },
    { name = "Daniel Lohmeier", email = "daniel.lohmeier@retoflow.de" },
    { name = "Jolando Marius Kisse", email = "jolando.kisse@uni-kassel.de" },
    { name = "Erik Prade", email = "erik.prade@iee.fraunhofer.de" }
]
description = "A pipeflow calculation tool that complements pandapower in the simulation of multi energy grids"
readme = "README.rst"
license = { file = "LICENSE" }
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    # Add the specific Python versions supported here, e.g.:
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "pandapower >= 3.1",
    "matplotlib",
    "shapely",
]
keywords = [
    "network", "analysis", "optimization", "automation", "grid", "energy", "engineering",
    "simulation", "pipeflow", "pandapipes", "gas"
]

[project.urls]
Homepage = "https://www.pandapipes.org"
Documentation = "https://pandapipes.readthedocs.io"
Source = "https://github.com/e2nIEE/pandapipes"
Repository = "https://github.com/e2nIEE/pandapipes.git"
Issues = "https://github.com/e2nIEE/pandapipes/issues"
Download = "https://pypi.org/project/pandapipes/#files"
Changelog = "https://github.com/e2nIEE/pandapipes/blob/develop/CHANGELOG.rst"

[project.optional-dependencies]
docs = ["numpydoc>=1.5.0", "sphinx>=5.3.0", "sphinx_rtd_theme>=1.1.1", "sphinxcontrib.bibtex>=2.5.0", "sphinx-pyproject"]
plotting = ["plotly", "python-igraph>=0.11.9,<0.12.0"]
test = ["pytest", "pytest-xdist", "pytest-split", "nbmake", "numba", "setuptools; python_version >= '3.12'"]
all = [
    "numpydoc>=1.5.0", "sphinx>=5.3.0", "sphinx_rtd_theme>=1.1.1", "sphinxcontrib.bibtex>=2.5.0", "sphinx-pyproject",
    "plotly", "igraph",
    "pytest", "pytest-xdist", "pytest-split", "nbmake", "numba","setuptools; python_version >= '3.12'"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["pandapipes*"]
